client/controlbar: Break up form into multiple chunks
To have radio selections with the same name in different tabs, we have
to break up the form. As they are entirely unrelated and never all
on-screen, this is more sane anyway.

Using a fieldset as a tab is somewhat nonsensical at this point, but
not invalid, so keep them for now.