client: Replace 'state_alter' with 'state_tweak' that doesn't call reload
When we do selections we ideally want to perform minor updates of interface
based on state. Instead of forcing no reloads, add a different function to
call in this case.

This is possibly also a hack, and would be better off with the reload function
deciding based on what's changed. But will do for now.