client: Don't trap users after errors, improve error logging
The final "throw err" meant that any next runs of the promise chain (e.g.
clicking a valid link after a 404) would just display the same error.

Remove it and replace with more useful console.error logging, so we see the
post-sourcemap version of the stacktrace.

The controlbar also stopped you from leaving a 404 page.