client/count: Show chapter counts for each book returned
Instead of having it as part of the texts display, show the chapter counts
here. This admittedly won't work if this becomes a more general count of things
page, but is probably not unreasonable for a summary.

Instead of wedging it into the query, re-use the chapter_start metadata option
to get a count of chapters