client/concordance: Count unique books ourselves, don't rely on metadata
The metadata is (now) returning metadata for all corpora items selected. This
is (probably) a good thing, since it inadvertantantly fixed a bug with rel.
freq. and means that we could push it out to it's own endpoint in the future,
so results get cached per-corpora.

Do our own counting, so we no longer rely on only books in the results being
returned.