client/page_text: Remember position in book last scrolled to
Once chapter_num gets populated, page was scrolling to chapter headings
whenever "quote" (e.g.) was selected, instead of preserving location.

Add an explicit record of where we are in the book, so that we return to that
on any other page alterations / reloads. Only override this when we chapter_num
changes post-page-load (i.e. a new one has been selected).