server/subset: Tokenless subsets should be [], not NULL
full_tokens aggregates all tokens in a subset. if there aren't any ensure
that it still returns an array.

This was only triggered by direct API calls, not the CLiC UI. The CLiC UI
requires left/right context, so there was always tokens, even if the node
was empty. The R clicclient OTOH does not request context.