flexiclic: Group algorithms by class
Instead of having a group for each type, have just "annotation" and
"algo", which we'll call algorithm class.

Shorten the querystring parameters to algo[%d] & annotation[%d], to
achieve this stop using the somewhat misguided attempts of shoehorning
nested args inside state.args() & make it a public function. This
relied on the parameter being an object, which isn't true any more.