client/controlbar: Operational flexiconc-path-chooser
Store all available paths in fc-all-paths, switch the contents of algo
in and out with this.

Use a set of radio controls to represent the path chooser, both to
contain the current path & allow the user to select the next.

Fork selects the hidden "next-path" option, clearing any subsequent
options.