page_promise: Don't clear alerts when tweaking
Flexiconc will tweak on a page-reload as a matter of course, to sync
the current path with the analysis tree. In doing so, it'll clear any
errors that appear.

By in large errors will be produced by reloads, not tweaks. Tweaking
won't cure the errors that a reload introduces, so we shouldn't remove
the messages.