util_flexiconc: page_state.nested_args -> util_flexiconc.renest_args
We need to apply nesting to args that come from fc-paths-all as well as
page_state. And it's bespoke functionality that doesn't really belong
here anyway.

Move page_state.nested_args() to a helper function, replace previous
uses by combining with page_state.all_args().

Note that as a result we stop supporting args("animals[cows][]"), but
this doesn't seem to be used anywhere anyway.