flexiclic: Support fetching packages at run-time
Add an async install_package_fn hook, use this to call out to micropip
for every depedency that algorithms ask for.

Modify manifest.js to be one big dict, use this to specify overrides to
the list of packages (e.g. instead of fetching "pyicu", fetch our built
wheel & the data wheel.