client/chosen_init: Use tomselect in FlexiConc
The allow-add-items hack doesn't make any sense, as if you try and add
"1" after adding "-1" there won't be no results.

Instead, add tomselect, use that instead for FlexiConc options.

This should be enough to ditch chosen-select for tomselect, but don't
do it yet in case there's other bugs.

We could wholesale replace chosen-select at this point, but it's UI for
single selects is a bit weird