flexiclic: Require confirmation to refresh table
Add an user error level of "confirm", which translates to showing a
refresh button atop the page content.

Add a concept of state_speculative_update, which the form update is the
main example of, add page_state.speculative to inform pages that the
current reload is speculative.

Finally, if speculative is true and we're trying to reload CLiC
data/annotations, throw the speculative error.