flexiconc: Add lineid-picker widgets
For fields named line_id or groups, add a custom widget that opens a
new CLiC window in an iframe, showing the path at that state.

Note that because it's in an iframe, this will proxy through to the
main page's webworker, and so there shouldn't be anything to refetch.

Add fc-select options for that table that adds checkboxes for the
relevant rows, and exposes a method to fetch them back again.

Finally, on OK read thes values back into the form.