flexiclic: Normalise opts/annotations before comparsion
Normalise opts/annotations before we try comparing them. This was
causing issues with spaCy to constantly fail equality check.

This may also be related to #31, which is still an issue. To help with
this further, re-arrange so we will definitely try and recreate
_flexiconc when missing, even if parameters haven't changed.

References: https://github.com/fau-klue/flexiconc/issues/31