path_util: Don't apply default to bool
A bool is going to be None/"on" before processing, where None is the
unchecked value. If a default is True, this will force the value to be
True regardless.

Ignore boolean defaults, and instead rely on the checkbox being
checked-by-default in this case.

References: https://github.com/fau-klue/flexiconc/issues/43