page_flexiconc: Update available algorithms after viewing a node
The available algorithms are now dependent on annotations being
available (e.g.), so getting available algorithms based on the root
node is no longer enough.

Return available algorithms as part of the compute_path metadata, and
use it to update the dropdowns after each query, along with a generic
update on init.

This means that the add algorithm dropdowns will be stale if a
speculative update fails (i.e. new annotations), but there isn't much
of a solution to that, bar dropping speculative updates.

Ideally the dropdown-updating would stay in controlbar_flexiconc and
data fed through the new_data callback, but that can be tidied up
later.